Renamed helpers tests to match src.
